Another major changes: 1. In Models: Remove all relationships between Courses, AcademicRecords, EnrollmentList, Lecturer, Schedule and Semester. 2. In Repositories: Remove the AssignCourseToStudent and GetCourseByStudent method in AdminRepository. Make some minor changes in CourseRepository. Remove the GetEnrollmentListByStudent and AddEnrollmentListToCourse in EnrollmentListRepository. Remove the UpdateStudentScore in LecturerRepository. Remove the GetCoursesByStudentId in StudentRepository. Also remove these methods in their respective Interface. 3. In CourseController: Make some changes to the method structures. 4. In ApplicationDbContext: Accomodate changes in the Models folder. 5. In Services: Remove the AssignCourseToStudent method in CourseService and make some changes to the remaining methods. Also remove the AssignCourseToStudent method in ICourseService. 6. in Account: Add a link that move the user to the Login page. 7. In Course: Add a "if" to check if Model is not null. Also make some minor changes. 8. Create a CreateCourse.html file and EditCourse.cshtml file. 9. Change the MapControllerRoute to the Login page instead of Register

+++ APDPAssignment/Repositories/CourseRepository.cs
@@ -20,7 +20,7 @@ namespace APDPAssignment.Repositories
             }
             catch
             {
-                return null;
+                return Enumerable.Empty<Course>();
             }
         }
 
@@ -68,10 +68,14 @@ namespace APDPAssignment.Repositories
         {
             try
             {
-                var course = _context.Course.Find(courseId);
-                _context.Course.Remove(course);
-                _context.SaveChanges();
-                return true;
+                var course = _context.Course.Find(courseId); 
+                if (course != null)
+                {
+                    _context.Course.Remove(course);
+                    _context.SaveChanges();
+                    return true;
+                }
+                return false;
             }
             catch
             {
